+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.2.0 - October 15, 2014
+
+* **BREAKING**: Removed extra filters
+* **CHANGE**: Take Nunjucks template engine in place of Swig to support Async rendering which make able to write Async filters
+
 ## v0.1.1 - September 30, 2014
 
 * Proof of concept.